+++ src/Components/quiz.js
@@ -1,0 +1,2 @@
+let quiz=document.getElementById("quiz");
+quiz.children[0]
