--- src/webpages/quiz.js
@@ -1,2 +1,0 @@
-let quiz=document.getElementById("quiz");
-quiz.children[0]
