--- src/webpages/spline.js
@@ -1,8 +1,0 @@
-import Spline from '@splinetool/react-spline';
-
-export default function SplineApp() {
-  return (
-    <Spline scene="https://prod.spline.design/xlEp3y2Z8j2P69Gq/scene.splinecode" />
-  );
-  
-}
